+++ DKImageManager/Data/Model/DKAsset.swift
@@ -137,8 +137,34 @@ public class DKAsset: NSObject {
 		}
 	}
 	
-	public func fetchAVAssetWithCompleteBlock(completeBlock: (avAsset: AVURLAsset?) -> Void) {
-		getImageManager().fetchAVAsset(self, completeBlock: completeBlock)
+	/**
+	Fetch an AVAsset with a completeBlock.
+	*/
+	public func fetchAVAssetWithCompleteBlock(completeBlock: (AVAsset: AVURLAsset?) -> Void) {
+		self.fetchAVAsset(nil, completeBlock: completeBlock)
+	}
+	
+	/**
+	Fetch an AVAsset with a completeBlock and PHVideoRequestOptions.
+	*/
+	public func fetchAVAsset(options: PHVideoRequestOptions?, completeBlock: (AVAsset: AVURLAsset?) -> Void) {
+		getImageManager().fetchAVAsset(self, options: options, completeBlock: completeBlock)
+	}
+	
+	/**
+	Sync fetch an AVAsset with a completeBlock and PHVideoRequestOptions.
+	*/
+	public func fetchAVAsset(sync: Bool, options: PHVideoRequestOptions?, completeBlock: (AVAsset: AVURLAsset?) -> Void) {
+		if sync {
+			let semaphore = dispatch_semaphore_create(0)
+			self.fetchAVAsset(nil, completeBlock: { (AVAsset) -> Void in
+				completeBlock(AVAsset: AVAsset)
+				dispatch_semaphore_signal(semaphore)
+			})
+			dispatch_semaphore_wait(semaphore, DISPATCH_TIME_FOREVER)
+		} else {
+			self.fetchAVAsset(nil, completeBlock: completeBlock)
+		}
 	}
 	
 }