+++ CHANGELOG.md
@@ -1,5 +1,51 @@
 # Change Log
 
+## [3.1.1](https://github.com/zhangao0086/DKImagePickerController/tree/3.1.1) (2016-03-18)
+
+[Full Changelog](https://github.com/zhangao0086/DKImagePickerController/compare/3.0.10...3.1.1)
+
+**Merged pull requests:**
+
+- Fixed an issue that may cause crash when user not authorized camera access.
+
+## [3.1.0](https://github.com/zhangao0086/DKImagePickerController/tree/3.1.0) (2016-03-17)
+
+[Full Changelog](https://github.com/zhangao0086/DKImagePickerController/compare/3.0.11...3.1.0)
+
+**Merged pull requests:**
+
+- Added support for custom camera.
+
+- Added support for UIDelegate.
+
+- Added a function to sync fetch an AVAsset.
+
+- Fixed an issue that may cause crashing when downloading image from iCloud.
+
+## [3.0.11](https://github.com/zhangao0086/DKImagePickerController/tree/3.0.11) (2016-02-27)
+
+[Full Changelog](https://github.com/zhangao0086/DKImagePickerController/compare/3.0.10...3.0.11)
+
+**Merged pull requests:**
+
+- Added a PHVideoRequestOptions API to fetch AVAsset.
+
+**Closed issues:**
+
+- Synchronous options for multiple video fetch [\#76](https://github.com/zhangao0086/DKImagePickerController/pull/76)
+
+## [3.0.10](https://github.com/zhangao0086/DKImagePickerController/tree/3.0.10) (2016-02-04)
+
+[Full Changelog](https://github.com/zhangao0086/DKImagePickerController/compare/3.0.9...3.0.10)
+
+**Merged pull requests:**
+
+- Added possibility to deselect all selected assets when showing a single instance picker.
+
+**Closed issues:**
+
+- Possibility to deselect assets when displaying picker for second time. [\#69](https://github.com/zhangao0086/DKImagePickerController/pull/69)
+
 ## [3.0.9](https://github.com/zhangao0086/DKImagePickerController/tree/3.0.9) (2016-01-29)
 
 [Full Changelog](https://github.com/zhangao0086/DKImagePickerController/compare/3.0.8...3.0.9)